Remove unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/dbcp/trunk@672592 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/commons/dbcp/datasources/PerUserPoolDataSource.java
@@ -55,7 +55,7 @@ public class PerUserPoolDataSource
 
     private int defaultMaxActive = GenericObjectPool.DEFAULT_MAX_ACTIVE;
     private int defaultMaxIdle = GenericObjectPool.DEFAULT_MAX_IDLE;
-    private int defaultMaxWait = (int)Math.min((long)Integer.MAX_VALUE,
+    private int defaultMaxWait = (int)Math.min(Integer.MAX_VALUE,
         GenericObjectPool.DEFAULT_MAX_WAIT);
     Map perUserDefaultAutoCommit = null;    
     Map perUserDefaultTransactionIsolation = null;
